[UI] code refactor
